# Change persistent reduction threshold to 32