Summary:

Increasing threshold for inductor multikernel flag from 16->32 can lead to significant performance gain. 

Example benchmark:
`# import torch
# import torch.nn.functional as F
# from triton.testing import do_bench
# import math


# num_buckets = 32
# max_distance=128

# def position_bias_softmax(scores, weight):
#     ##### Calculate position bias
#     context_position = torch.arange(2048, dtype=torch.long, device="cuda")[:, None]
#     memory_position = torch.arange(2048, dtype=torch.long, device="cuda")[None, :]
#     relative_position = memory_position - context_position  # shape (query_length, key_length)
#     relative_buckets = 0
#     relative_position = -torch.min(relative_position, torch.zeros_like(relative_position))
#     max_exact = num_buckets // 2
#     is_small = relative_position < max_exact
#     relative_position_if_large = max_exact + (
#         torch.log(relative_position.float() / max_exact)
#         / math.log(max_distance / max_exact)
#         * (num_buckets - max_exact)
#     ).to(torch.long)
#     relative_position_if_large = torch.min(
#         relative_position_if_large, torch.full_like(relative_position_if_large, num_buckets - 1)
#     )

#     relative_buckets += torch.where(is_small, relative_position, relative_position_if_large)

#     ##### Embedding lookup
#     position_bias = F.embedding(relative_buckets, weight)
#     position_bias = position_bias.permute([2, 0, 1]).unsqueeze(0)


#     ##### Apply position bias and softmax
#     scores = scores + position_bias

#     return F.softmax(scores.float(), dim=-1).to(torch.float16)



# scores = torch.randn(1, 8, 2048, 2048, device="cuda", dtype=torch.float16)
# weight = torch.randn(32, 8, device="cuda")
# output = position_bias_softmax(scores, weight=weight)

# compiled = torch.compile(position_bias_softmax)
# # compiled(scores)
# # compiled(scores, pw_bias=True)
# compiled_output = compiled(scores, weight=weight)
# gb = 2 * scores.element_size() * scores.numel() / 1e9
# # sec = do_bench(lambda: compiled(scores)) / 1e3
# # print(f"no bias gb/s: {gb/sec}")
# # sec = do_bench(lambda: compiled(scores, pw_bias=True)) / 1e3
# # print(f"pointwise bias gb/s: {gb/sec}")
# sec = do_bench(lambda: compiled(scores, weight=weight)) / 1e3

# print("Allclose: ", torch.allclose(output, compiled_output, atol=1e-3, rtol=1e-3))
# print(f"weighted bias gb/s: {gb/sec}")


import torch
import torch.nn.functional as F
from triton.testing import do_bench
from torch._inductor import config as inductor_config
import math

def position_bias_softmax(scores, weight=None, pw_bias=False):
    scores = scores.to(torch.float32)
    context_position = torch.arange(2048, dtype=torch.long, device="cuda")[:, None]
    memory_position = torch.arange(2048, dtype=torch.long, device="cuda")[None, :]
    relative_position = memory_position - context_position  # shape (query_length, key_length)
    relative_buckets = 0
    num_buckets=32
    max_distance=128
    relative_position = -torch.min(relative_position, torch.zeros_like(relative_position))
    max_exact = num_buckets // 2
    is_small = relative_position < max_exact
    relative_position_if_large = max_exact + (
        torch.log(relative_position.float() / max_exact)
        / math.log(max_distance / max_exact)
        * (num_buckets - max_exact)
    ).to(torch.long)
    relative_position_if_large = torch.min(
        relative_position_if_large, torch.full_like(relative_position_if_large, num_buckets - 1)
    )

    relative_buckets += torch.where(is_small, relative_position, relative_position_if_large)
    # relative_buckets = relative_position_if_large.clamp(0, 31)
    values = F.embedding(relative_buckets, weight)
    values = values.permute([2, 0, 1]).unsqueeze(0) 
    scores = scores + values

    return F.softmax(scores, dim=-1).to(torch.float16)


scores = torch.randn(8, 2048, 2048, device="cuda", dtype=torch.float16)
weight = torch.randn(32, 1, device="cuda")
position_bias_softmax(scores, weight)
compiled = torch.compile(position_bias_softmax)
# compiled(scores)
# compiled(scores, pw_bias=True)
compiled(scores, weight=weight)
gb = 2 * scores.element_size() * scores.numel() / 1e9
# sec = do_bench(lambda: compiled(scores)) / 1e3
# print(f"no bias gb/s: {gb/sec}")
# sec = do_bench(lambda: compiled(scores, pw_bias=True)) / 1e3
# print(f"pointwise bias gb/s: {gb/sec}")
sec = do_bench(lambda: compiled(scores, weight=weight)) / 1e3
print(f"weighted bias gb/s: {gb/sec}")
`

With this change: gb/s: 987.0799446648006
Baseline: gb/s: 693.3391918370983